COMPLETE map_generators.py modularization - MODULARIZATION FINISHED
- Created map_locations/map_generators.py with 5 functions (1,161 lines)
- Reduced core.py from 1,213 to 72 lines (-1,134 lines)
- Updated core.py to import from map_generators module
- All backward compatibility maintained
- Testing: All map generation functions work correctly

MODULARIZATION COMPLETE: 100% (6/6 modules)
categories.py - utils.py - location_filters.py
popup_generators.py - exporters.py - map_generators.py

FINAL RESULTS:
- Original core.py: 2,552 lines
- Final core.py: 72 lines
- Total reduction: 2,480 lines (97% smaller)
- Modules created: 6 focused, maintainable modules
- All functionality preserved and tested

Mission accomplished according to plan
